requirements.txt
# Core FastAPI
fastapi==0.109.0
uvicorn==0.27.0

# Database
sqlalchemy==2.0.25

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# AI Integration
openai==1.12.0

# Voice
SpeechRecognition==3.10.1
gTTS==2.5.1
playsound==1.2.2
pywin32==306

# HTTP Requests
requests==2.31.0
